requirements.txt
google-api-core==2.25.1
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-cloud-bigquery==3.38.0
google-cloud-core==2.4.3
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.75.0
grpcio-status==1.75.0
idna==3.10
numpy==2.3.3
oauthlib==3.3.1
packaging==25.0
pandas==2.3.2
pandas-gbq==0.29.2
proto-plus==1.26.1
protobuf==6.32.1
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydata-google-auth==1.9.1
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
